Add books list feature as example
