+++ client/src/components/book/index.js
@@ -1,0 +1,3 @@
+import List from './List';
+
+export {List};
