+++ client/src/index.js
@@ -12,9 +12,12 @@ import 'font-awesome/css/font-awesome.css';
 import registerServiceWorker from './registerServiceWorker';
 // Import your reducers and routes here
 import Welcome from './Welcome';
+import Theme from './Theme';
+import book from './reducers/book/';
+import bookRoutes from './routes/book';
 
 const store = createStore(
-  combineReducers({routing, form, /* Add your reducers here */}),
+  combineReducers({routing, form, book, /* Add your reducers here */}),
   applyMiddleware(thunk),
 );
 
@@ -22,13 +25,16 @@ const history = syncHistoryWithStore(createBrowserHistory(), store);
 
 ReactDom.render(
   <Provider store={store}>
-    <Router history={history}>
-      <Switch>
-        <Route path="/" component={Welcome} strict={true} exact={true}/>
-        {/* Add your routes here */}
-        <Route render={() => <h1>Not Found</h1>}/>
-      </Switch>
-    </Router>
+    <Theme>
+      <Router history={history}>
+        <Switch>
+          <Route path="/" component={Welcome} strict={true} exact={true}/>
+          {bookRoutes}
+          {/* Add your routes here */}
+          <Route render={() => <h1>Not Found</h1>}/>
+        </Switch>
+      </Router>
+    </Theme>
   </Provider>,
   document.getElementById('root')
 );
