+++ client/src/reducers/book/index.js
@@ -1,0 +1,4 @@
+import {combineReducers} from 'redux';
+import list from './list';
+
+export default combineReducers({list});
