+++ client/src/routes/book.js
@@ -1,0 +1,7 @@
+import React from 'react';
+import {Route} from 'react-router-dom';
+import {List} from '../components/book/';
+
+export default [
+  <Route path='/books/:page?' component={List} strict={true} key='list'/>,
+];
